Implement .NET API for Connector
The API is C#-friendly. Thanks to the P/Invoke technology it works in Windows, Linux and Mac OS X. It works with .NET Framework 3.5 and higher and .NET Standard 1.1 (which implies mono and .NET Core 1.0 and higher).

I have also added a couple of examples: Simple, Mixed and Object.